+++ Products/CMFPlone/utils.py
@@ -709,6 +709,11 @@ def getQuality():
     return QUALITY_DEFAULT
 
 
+def getRetinaScales():
+    from plone.namedfile.utils import getRetinaScales as func
+    return func()
+
+
 def getSiteLogo(site=None):
     from Products.CMFPlone.interfaces import ISiteSchema
     from plone.formwidget.namedfile.converter import b64decode_file